third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/edit-dom-actions-2-expected.txt

Tests that user can mutate DOM by means of elements panel.


Running: testSetUp

Running: testSetAttribute
==== before ====
- <div id="testSetAttribute">
      <div foo="attribute value" id="node-to-set-attribute"></div>
  </div>
==== after ====
- <div id="testSetAttribute">
      <div id="node-to-set-attribute" bar="edited attribute"></div>
  </div>

Running: testSetScriptableAttribute
==== before ====
- <div id="testSetScriptableAttribute">
      <div onclick="alert(1)" id="node-to-set-scriptable-attribute"></div>
  </div>
==== after ====
- <div id="testSetScriptableAttribute">
      <div onclick="alert(2)" id="node-to-set-scriptable-attribute"></div>
  </div>

Running: testRemoveAttribute
==== before ====
- <div id="testRemoveAttribute">
      <div foo="attribute value" id="node-to-remove-attribute"></div>
  </div>
==== after ====
- <div id="testRemoveAttribute">
      <div id="node-to-remove-attribute"></div>
  </div>

Running: testAddAttribute
==== before ====
- <div id="testAddAttribute">
      <div id="node-to-add-attribute"></div>
  </div>
==== after ====
- <div id="testAddAttribute">
      <div id="node-to-add-attribute" newattr="new-value"></div>
  </div>

Running: testAddAttributeUnquotedValue
==== before ====
- <div id="testAddAttributeUnquotedValue">
      <div id="node-to-add-attribute-unquoted-value"></div>
  </div>
==== after ====
- <div id="testAddAttributeUnquotedValue">
      <div id="node-to-add-attribute-unquoted-value" newattr="unquotedValue"></div>
  </div>
